+++ R/is_num_fac.R
@@ -1,11 +1,14 @@
 #' @title Check whether a factor has numeric levels only
 #' @name is_num_fac
-#' @description This function checks whether a factor has only numeric or
-#'                any non-numeric factor levels.
+#' @description \code{is_num_fac()} checks whether a factor has only numeric or
+#'    any non-numeric factor levels, while \code{is_num_chr()} checks whether
+#'    a character vector has only numeric strings.
 #'
-#' @param x A \code{\link{factor}}.
-#' @return Logical, \code{TRUE} if factor has numeric factor levels only,
-#'           \code{FALSE} otherwise.
+#' @param x A factor for \code{is_num_fac()} and a character vector for
+#'    \code{is_num_chr()}
+#'
+#' @return Logical, \code{TRUE} if factor has numeric factor levels only, or
+#'    if character vector has numeric strings only, \code{FALSE} otherwise.
 #'
 #' @examples
 #' # numeric factor levels
@@ -20,8 +23,18 @@
 #' f3 <- factor(c("Justus", "Bob", "Peter"))
 #' is_num_fac(f3)
 #'
+#' is_num_chr(c("a", "1"))
+#' is_num_chr(c("2", "1"))
+#'
 #' @export
 is_num_fac <- function(x) {
   # check if we have numeric levels
   !anyNA(suppressWarnings(as.numeric(levels(x))))
+}
+
+#' @rdname is_num_fac
+#' @export
+is_num_chr <- function(x) {
+  # check if we have numeric bvalues
+  !anyNA(suppressWarnings(as.numeric(x)))
 }