+++ R/select_helpers.R
@@ -36,9 +36,15 @@ obj_has_rownames <- function(x) {
 }
 
 #' @importFrom dplyr select
-add_cols <- function(data, ..., .after = 1) {
+add_cols <- function(data, ..., .after = 1, .before = NULL) {
   if (is.character(.after))
     .after <- which(colnames(data) == .after)
+
+  if (!is.null(.before) && is.character(.before))
+    .after <- which(colnames(data) == .before) - 1
+
+  if (!is.null(.before) && is.numeric(.before))
+    .after <- .before - 1
 
   dat <- data.frame(..., stringsAsFactors = FALSE)